Remove Actor class
