+++ packages/core/src/ObservableActorRef.ts
@@ -1,16 +1,21 @@
 import { EventObject, ActorRef, Behavior, SCXML, ActorContext } from './types';
-import { Actor } from './actor';
 import { symbolObservable, toSCXMLEvent } from './utils';
+import { Mailbox } from './Mailbox';
+import { LifecycleSignal, startSignal, stopSignal } from './actors';
+
+const nullSubscription = {
+  unsubscribe: () => void 0
+};
 
 export class ObservableActorRef<TEvent extends EventObject, TEmitted>
   implements ActorRef<TEvent, TEmitted> {
   private current: TEmitted;
   public deferred = true;
   private context: ActorContext<TEvent, TEmitted>;
-  private actor: Actor<TEvent, TEmitted>;
   public name: string;
+  private mailbox = new Mailbox(this._process.bind(this));
 
-  constructor(behavior: Behavior<TEvent, TEmitted>, name: string) {
+  constructor(public behavior: Behavior<TEvent, TEmitted>, name: string) {
     this.name = name;
     this.context = {
       self: this,
@@ -18,24 +23,25 @@ export class ObservableActorRef<TEvent extends EventObject, TEmitted>
       observers: new Set(),
       _event: toSCXMLEvent({ type: 'xstate.init' }) as SCXML.Event<TEvent> // TODO: fix
     };
-    this.actor = new Actor(behavior, name, this.context);
-    this.current = this.actor.current;
+    this.current = this.behavior.initialState;
   }
   public start() {
     this.deferred = false;
-    this.actor.start();
+    this.mailbox.enqueue(startSignal);
+    this.mailbox.start();
 
     return this;
   }
   public stop() {
-    this.actor.stop();
+    this.mailbox.clear();
+    this.mailbox.enqueue(stopSignal);
     return this;
   }
   public subscribe(observer) {
-    return this.actor.subscribe(observer);
+    return this.behavior.subscribe?.(observer) ?? nullSubscription;
   }
   public send(event) {
-    this.actor.receive(event);
+    this.receive(event);
   }
   public toJSON() {
     return {
@@ -44,9 +50,26 @@ export class ObservableActorRef<TEvent extends EventObject, TEmitted>
     };
   }
   public getSnapshot() {
-    return this.actor.current;
+    return this.current;
   }
   public [symbolObservable]() {
     return this;
+  }
+  public receive(event: TEvent | LifecycleSignal) {
+    this.mailbox.enqueue(event);
+  }
+  private _process(event: TEvent | LifecycleSignal) {
+    this.context._event =
+      typeof event.type !== 'string'
+        ? (event as LifecycleSignal)
+        : toSCXMLEvent(event as TEvent);
+
+    this.current = this.behavior.transition(
+      this.current,
+      typeof this.context._event.type !== 'string'
+        ? (this.context._event as LifecycleSignal)
+        : (this.context._event as SCXML.Event<TEvent>).data,
+      this.context
+    );
   }
 }
